+++ pkg/agent/resources.go
@@ -6,6 +6,7 @@ import (
 	"math"
 
 	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
 	"resourcemanagement.controlplane/pkg/ctlplaneapi"
 )
 
@@ -64,12 +65,35 @@ func GetDeletePodRequest(pod *corev1.Pod) *ctlplaneapi.DeletePodRequest {
 
 	return deletePodRequest
 }
-
+func addContainerMemoryToPod(cInfo *ctlplaneapi.ContainerInfo,
+	podRequestedMemory *resource.Quantity,
+	podLimitMemory *resource.Quantity) error {
+	rm := resource.Quantity{}
+	err := rm.Unmarshal(cInfo.Resources.RequestedMemory)
+	var zero resource.Quantity
+	if err != nil {
+		return err
+	}
+	if rm.Cmp(zero) < 0 {
+		return fmt.Errorf("mem request: %w", ErrCountingOverflow)
+	}
+	podRequestedMemory.Add(rm)
+	lm := resource.Quantity{}
+	err = lm.Unmarshal(cInfo.Resources.LimitMemory)
+	if err != nil {
+		return err
+	}
+	if lm.Cmp(zero) < 0 {
+		return fmt.Errorf("mem limit: %w", ErrCountingOverflow)
+	}
+	podLimitMemory.Add(lm)
+	return nil
+}
 func createPodResources(pod *corev1.Pod) ([]*ctlplaneapi.ContainerInfo, *ctlplaneapi.ResourceInfo, error) {
 	var podRequestedCpus int32
 	var podLimitCpus int32
-	var podRequestedMemory int32
-	var podLimitMemory int32
+	var podRequestedMemory resource.Quantity
+	var podLimitMemory resource.Quantity
 
 	containerInfo := make([]*ctlplaneapi.ContainerInfo, 0)
 
@@ -90,23 +114,26 @@ func createPodResources(pod *corev1.Pod) ([]*ctlplaneapi.ContainerInfo, *ctlplan
 		if podLimitCpus < 0 {
 			return containerInfo, nil, fmt.Errorf("cpus limit: %w", ErrCountingOverflow)
 		}
-		podRequestedMemory += cInfo.Resources.RequestedMemory
-		if podRequestedMemory < 0 {
-			return containerInfo, nil, fmt.Errorf("mem request: %w", ErrCountingOverflow)
-		}
-		podLimitMemory += cInfo.Resources.LimitMemory
-		if podLimitMemory < 0 {
-			return containerInfo, nil, fmt.Errorf("mem limit: %w", ErrCountingOverflow)
-		}
 
+		err = addContainerMemoryToPod(cInfo, &podRequestedMemory, &podLimitMemory)
+		if err != nil {
+			return []*ctlplaneapi.ContainerInfo{}, nil, err
+		}
 		containerInfo = append(containerInfo, cInfo)
 	}
-
+	rm, err := podRequestedMemory.Marshal()
+	if err != nil {
+		return containerInfo, nil, err
+	}
+	lm, err := podLimitMemory.Marshal()
+	if err != nil {
+		return containerInfo, nil, err
+	}
 	resourceInfo := &ctlplaneapi.ResourceInfo{
 		RequestedCpus:   podRequestedCpus,
 		LimitCpus:       podLimitCpus,
-		RequestedMemory: podRequestedMemory,
-		LimitMemory:     podLimitMemory,
+		RequestedMemory: rm,
+		LimitMemory:     lm,
 	}
 
 	return containerInfo, resourceInfo, nil
@@ -135,22 +162,20 @@ func getContainerInfo(container *corev1.Container) (*ctlplaneapi.ContainerInfo, 
 	return containerInfo, nil
 }
 
-func getContainerResources(resourceList corev1.ResourceList) (int32, int32, error) {
+func getContainerResources(resourceList corev1.ResourceList) (int32, []byte, error) {
 	cpusQuantity := resourceList.Cpu()
 	cpus, representable := cpusQuantity.AsInt64()
 
 	if !representable || cpus > math.MaxInt32 || cpus < 0 {
-		return 0, 0, fmt.Errorf("cpu quantity %v: %w", cpusQuantity, ErrNotRepresentable)
+		return 0, nil, fmt.Errorf("cpu quantity %v: %w", cpusQuantity, ErrNotRepresentable)
 	}
 
 	memoryQuantity := resourceList.Memory()
-	memory, representable := memoryQuantity.AsInt64()
-
-	if !representable || memory > math.MaxInt32 || memory < 0 {
-		return 0, 0, fmt.Errorf("mem quantity %v: %w", memoryQuantity, ErrNotRepresentable)
+	memory, err := memoryQuantity.Marshal()
+	if err != nil {
+		return 0, nil, err
 	}
-
-	return int32(cpus), int32(memory), nil
+	return int32(cpus), memory, nil
 }
 
 func getContainerID(name string, pod *corev1.Pod) string {