+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.1[24.01.2022]
+### Bugfixes
+- numa-namespace policies handles correclty more namespaces than numa zones by grouping them 
+- report error when runtime agent configuration does not match the system runtime
+- correct representation of memory resources
+- correct cgroup v1 path validation 
+
 ## 0.1 [02.12.2022]
 ### Added
 - Add support for cgroups v2